Boss AI Completion, Weapon Pickups, Healthbars
- Weapon inventory completely implemented
- Boss AI completed
- Collectible Weapon Prefabs Added (Revolver, Shotgun, and Crowbar)
- Bullets now leave trail effect

+++ Assets/Scripts/PlayerHealth.cs
@@ -3,8 +3,8 @@ using UnityEngine.SceneManagement;
 
 public class PlayerHealth : MonoBehaviour
 {
-    public int maxHealth = 3;
-    public int currentHealth;
+    public float maxHealth = 3;
+    public float currentHealth;
 
     void Start()
     {
@@ -23,7 +23,7 @@ public class PlayerHealth : MonoBehaviour
         }
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
         Debug.Log("Player Health: " + currentHealth);
@@ -34,7 +34,7 @@ public class PlayerHealth : MonoBehaviour
         }
     }
 
-    public int GetCurrentHealth()
+    public float GetCurrentHealth()
     {
         return currentHealth;
     }
